Blog/requirement.txt
pymysql
django==1.11.9
aliyun-python-sdk-core
django-simple-captcha
pillow
baidu-aip
django-mdeditor
django-crispy-forms>=1.6.0
django-import-export>=0.5.1
django-reversion>=2.0.0
django-formtools==1.0
future==0.15.2
httplib2==0.9.2
six==1.10.0
markdown
django-formtools==2.1
django-qiniu-storge
fuzzywuzzy
python-Levenshtein
django-contrib-comments
dj-pagination